+++ lib/machine/stack/list/debug.h
@@ -9,10 +9,10 @@
 
 void stack_l_display(evm_stack_l_p sl);
 
-evm_stack_l_p stack_l_init_immed_variadic(int n, va_list *args);
+evm_stack_l_p stack_l_init_immed_variadic(uint64_t n, va_list *args);
 
-bool stack_l_test_immed(evm_stack_l_p sl, int n, ...);
-bool stack_l_test_variadic(evm_stack_l_p sl, int n, va_list *args);
+bool stack_l_test_immed(evm_stack_l_p sl, uint64_t n, ...);
+bool stack_l_test_variadic(evm_stack_l_p sl, uint64_t n, va_list *args);
 
 #endif
